feat(chat): introduce auditMode, room deletions on leave and remove unique name requirement from rooms (#794)
feat(chat): delete room when all participants have left
fix(chat): room deletion not deleting associated messages
refactor(chat): allow rooms with non-unique names
feat(chat): introduce audit mode to keep chats and rooms after deletion
feat(chat): add participant log to keep a history of who joins or leaves a room
refactor(grpc-sdk): add missing types in query
fix(chat): remove "any" from query
fix(authorization): remove "any" from query
chore: codefactor issues

+++ modules/chat/src/routes/index.ts
@@ -69,16 +69,19 @@ export class ChatRoutes {
     } catch (e) {
       throw new GrpcError(status.INTERNAL, (e as Error).message);
     }
-    const roomExists = await ChatRoom.getInstance()
-      .findOne({ name: roomName })
-      .catch((e: Error) => {
-        throw new GrpcError(status.INTERNAL, e.message);
-      });
-    if (!isNil(roomExists)) {
-      throw new GrpcError(status.ALREADY_EXISTS, `Room ${roomName} already exists`);
-    }
     let room;
-    const query: Query<ChatRoom> = { name: roomName, participants: [user._id] };
+    const query: Query<ChatRoom> = {
+      name: roomName,
+      creator: user._id,
+      participants: [user._id],
+      participantsLog: [
+        {
+          user: user._id,
+          action: 'create',
+          timestamp: new Date(new Date().getUTCMilliseconds()),
+        },
+      ],
+    };
     const config = await this.grpcSdk.config.get('chat');
     if (config.explicit_room_joins.enabled) {
       room = await ChatRoom.getInstance()
@@ -100,6 +103,18 @@ export class ChatRoutes {
       });
     } else {
       query['participants'] = Array.from(new Set([user._id, ...users]));
+      query['participantsLog'] = [
+        {
+          user: user._id,
+          action: 'create',
+          timestamp: new Date(new Date().getUTCMilliseconds()),
+        },
+        ...users.map((userId: string) => ({
+          user: userId,
+          action: 'join' as 'join',
+          timestamp: new Date(new Date().getUTCMilliseconds()),
+        })),
+      ];
       room = await ChatRoom.getInstance()
         .create(query)
         .catch((e: Error) => {
@@ -108,7 +123,7 @@ export class ChatRoutes {
     }
     this.grpcSdk.bus?.publish(
       'chat:create:ChatRoom',
-      JSON.stringify({ name: roomName, participants: room.participants }),
+      JSON.stringify({ _id: room._id, name: roomName, participants: room.participants }),
     );
     return { roomId: room._id };
   }
@@ -129,6 +144,9 @@ export class ChatRoutes {
         throw new GrpcError(status.INTERNAL, e.message);
       },
     );
+    if (room.deleted) {
+      throw new GrpcError(status.NOT_FOUND, "Room doesn't exist");
+    }
     try {
       usersToBeAdded = await validateUsersInput(this.grpcSdk, users);
     } catch (e) {
@@ -162,6 +180,14 @@ export class ChatRoutes {
       await ChatRoom.getInstance()
         .findByIdAndUpdate(room._id, {
           participants: Array.from(new Set([...room.participants, ...users])),
+          participantsLog: [
+            ...room.participantsLog,
+            ...users.map((userId: string) => ({
+              user: userId,
+              action: 'join' as 'join',
+              timestamp: new Date(new Date().getUTCMilliseconds()),
+            })),
+          ],
         })
         .catch((e: Error) => {
           throw new GrpcError(status.INTERNAL, e.message);
@@ -180,15 +206,36 @@ export class ChatRoutes {
         throw new GrpcError(status.INTERNAL, e.message);
       },
     );
+    if (room.deleted) {
+      throw new GrpcError(status.NOT_FOUND, "Room doesn't exist");
+    }
 
     const index = room.participants.indexOf(user._id);
     if (index > -1) {
       room.participants.splice(index, 1);
+      room.participantsLog.push({
+        user: user._id,
+        action: 'leave' as 'leave',
+        timestamp: new Date(new Date().getUTCMilliseconds()),
+      });
       if (
-        room.participants.length === 1 &&
-        ConfigController.getInstance().config.deleteEmptyRooms
+        (room.participants.length === 1 &&
+          ConfigController.getInstance().config.deleteEmptyRooms) ||
+        room.participants.length === 0
       ) {
-        await ChatRoom.getInstance().deleteOne({ _id: room._id });
+        if (ConfigController.getInstance().config.auditMode) {
+          await ChatRoom.getInstance().findByIdAndUpdate(room._id, {
+            ...room,
+            deleted: true,
+          });
+          await ChatMessage.getInstance().updateMany(
+            { room: room._id },
+            { deleted: true },
+          );
+        } else {
+          await ChatRoom.getInstance().deleteOne({ _id: room._id });
+          await ChatMessage.getInstance().deleteMany({ room: room._id });
+        }
         this.grpcSdk.bus?.publish('chat:deleteRoom:ChatRoom', JSON.stringify({ roomId }));
       } else {
         await ChatRoom.getInstance()
@@ -213,11 +260,14 @@ export class ChatRoutes {
     let countPromise;
     if (isNil(roomId)) {
       const rooms = await ChatRoom.getInstance()
-        .findMany({ participants: user._id })
+        .findMany({ participants: user._id, deleted: false })
         .catch((e: Error) => {
           throw new GrpcError(status.INTERNAL, e.message);
         });
-      const query = { room: { $in: rooms.map((room: ChatRoom) => room._id) } };
+      const query = {
+        room: { $in: rooms.map((room: ChatRoom) => room._id) },
+        deleted: false,
+      };
       messagesPromise = ChatMessage.getInstance().findMany(
         query,
         undefined,
@@ -227,13 +277,19 @@ export class ChatRoutes {
       );
       countPromise = ChatMessage.getInstance().countDocuments(query);
     } else {
-      await this.fetchAndValidateRoomById(roomId, user._id).catch((e: Error) => {
-        throw new GrpcError(status.INTERNAL, e.message);
-      });
+      const room = await this.fetchAndValidateRoomById(roomId, user._id).catch(
+        (e: Error) => {
+          throw new GrpcError(status.INTERNAL, e.message);
+        },
+      );
+      if (room.deleted) {
+        throw new GrpcError(status.NOT_FOUND, "Room doesn't exist");
+      }
       messagesPromise = ChatMessage.getInstance()
         .findMany(
           {
             room: roomId,
+            deleted: false,
           },
           undefined,
           skip,
@@ -243,7 +299,10 @@ export class ChatRoutes {
         .catch((e: Error) => {
           throw new GrpcError(status.INTERNAL, e.message);
         });
-      countPromise = ChatMessage.getInstance().countDocuments({ room: roomId });
+      countPromise = ChatMessage.getInstance().countDocuments({
+        room: roomId,
+        deleted: false,
+      });
     }
 
     const [messages, count] = await Promise.all([messagesPromise, countPromise]).catch(
@@ -258,7 +317,7 @@ export class ChatRoutes {
     const { id } = call.request.params;
     const { user } = call.request.context;
     const message = await ChatMessage.getInstance()
-      .findOne({ _id: id }, undefined, ['room'])
+      .findOne({ _id: id, deleted: false }, undefined, ['room'])
       .catch((e: Error) => {
         throw new GrpcError(status.INTERNAL, e.message);
       });
@@ -275,12 +334,19 @@ export class ChatRoutes {
     const { skip, limit, populate } = call.request.params;
     const { user } = call.request.context;
     const rooms = await ChatRoom.getInstance()
-      .findMany({ participants: user._id }, undefined, skip, limit, undefined, populate)
+      .findMany(
+        { participants: user._id, deleted: false },
+        undefined,
+        skip,
+        limit,
+        undefined,
+        populate,
+      )
       .catch((e: Error) => {
         throw new GrpcError(status.INTERNAL, e.message);
       });
     const roomsCount = await ChatRoom.getInstance()
-      .countDocuments({ participants: user._id })
+      .countDocuments({ participants: user._id, deleted: false })
       .catch((e: Error) => {
         throw new GrpcError(status.INTERNAL, e.message);
       });
@@ -291,7 +357,7 @@ export class ChatRoutes {
     const { id, populate } = call.request.params;
     const { user } = call.request.context;
     const room = await ChatRoom.getInstance()
-      .findOne({ _id: id, participants: user._id }, undefined, populate)
+      .findOne({ _id: id, participants: user._id, deleted: false }, undefined, populate)
       .catch((e: Error) => {
         throw new GrpcError(status.INTERNAL, e.message);
       });
@@ -308,7 +374,7 @@ export class ChatRoutes {
     const { messageId } = call.request.params;
     const { user } = call.request.context;
     const message = await ChatMessage.getInstance()
-      .findOne({ _id: messageId })
+      .findOne({ _id: messageId, deleted: false })
       .catch((e: Error) => {
         throw new GrpcError(status.INTERNAL, e.message);
       });
@@ -318,11 +384,20 @@ export class ChatRoutes {
         "Message does not exist or you don't have access",
       );
     }
-    await ChatMessage.getInstance()
-      .deleteOne({ _id: messageId })
-      .catch((e: Error) => {
-        throw new GrpcError(status.INTERNAL, e.message);
-      });
+    if (ConfigController.getInstance().config.auditMode) {
+      await ChatMessage.getInstance()
+        .findByIdAndUpdate(messageId, { deleted: true })
+        .catch((e: Error) => {
+          throw new GrpcError(status.INTERNAL, e.message);
+        });
+    } else {
+      await ChatMessage.getInstance()
+        .deleteOne({ _id: messageId })
+        .catch((e: Error) => {
+          throw new GrpcError(status.INTERNAL, e.message);
+        });
+    }
+
     this.grpcSdk.bus?.publish('chat:delete:ChatMessage', JSON.stringify(messageId));
     return 'Message deleted successfully';
   }
@@ -331,7 +406,7 @@ export class ChatRoutes {
     const { messageId, newMessage } = call.request.params;
     const { user } = call.request.context;
     const message: ChatMessage | null = await ChatMessage.getInstance()
-      .findOne({ _id: messageId })
+      .findOne({ _id: messageId, deleted: false })
       .catch((e: Error) => {
         throw new GrpcError(status.INTERNAL, e.message);
       });
@@ -363,7 +438,7 @@ export class ChatRoutes {
   async onMessage(call: ParsedSocketRequest): Promise<UnparsedSocketResponse> {
     const { user } = call.request.context;
     const [roomId, message] = call.request.params;
-    const room = await ChatRoom.getInstance().findOne({ _id: roomId });
+    const room = await ChatRoom.getInstance().findOne({ _id: roomId, deleted: false });
 
     if (isNil(room) || !room.participants.includes(user._id)) {
       throw new GrpcError(
@@ -389,7 +464,7 @@ export class ChatRoutes {
   async onMessagesRead(call: ParsedSocketRequest): Promise<UnparsedSocketResponse> {
     const user: User = call.request.context.user;
     const [roomId] = call.request.params;
-    const room = await ChatRoom.getInstance().findOne({ _id: roomId });
+    const room = await ChatRoom.getInstance().findOne({ _id: roomId, deleted: false });
     if (isNil(room) || !(room.participants as string[]).includes(user._id)) {
       throw new GrpcError(
         status.INVALID_ARGUMENT,
